backend/requirements.txt
flask
flask-cors
joblib
pandas
scikit-learn==1.5.1
ucimlrepo
google-genai
python-dotenv
gunicorn
